HOUSNAV-225: initial multi dwelling 9.9.9 json
