+++ packages/data/tsconfig.json
@@ -1,4 +1,4 @@
 {
   "extends": "@repo/typescript-config/react-library.json",
-  "include": ["src", "json", "vite.config.mts"]
+  "include": ["src", "json", "vite.config.mts", "utils"]
 }
